backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-cleanup==9.0.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.8.2
frozenlist==1.5.0
geoip2==4.8.0
idna==3.10
maxminddb==2.6.2
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.3.0
typing_extensions==4.12.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
yarl==1.17.1
